Fix conditionals in get_storage_id
The checks for missing storage account ids would never fire as 'null' is
not a valid type in bash.

This resulted in an empty value for 'ids' being returned causing
subsequent errors in the 'az' command (and possibly elsewhere).

Update these to the standard bash '-z' check. Also remove the check for
null in functions.sh as this case is covered within get_storage_id.