Exit with error if no launchpad found
In the case where no matching environment is found or the user doesn't
have enough permissions, we should exit with non zero. An example use
case is where I mistype or otherwise forget to set my environment, rover
will output the "There is no remote state for ..." message, but in a
pipeline can look like it has passed.